requirements.txt
requests==2.28.1
python-dotenv==0.19.2
py==1.11.0
pytest==7.1.3
